# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Demo API to fetch Ethereum smart-contract info and ERC20 metadata using Infura",
        "title": "eth-contract-info-api Contract API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/contracts/erc20": {
            "get": {
                "description": "Fetch ERC20 name, symbol, decimals, and totalSupply using eth_call via Infura. Requires an ERC20 contract address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get ERC20 token metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ERC20 contract address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "latest",
                        "description": "Block tag",
                        "name": "blockTag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ERC20MetadataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/contracts/info": {
            "get": {
                "description": "Fetch deployed bytecode information for an Ethereum address using Infura. Contract address recommended; wallet address returns no code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Get smart contract info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethereum address (contract address recommended)",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "latest",
                        "description": "Block tag",
                        "name": "blockTag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ContractInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "contract.ContractInfoResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "blockTag": {
                    "type": "string"
                },
                "bytecodePreview": {
                    "type": "string"
                },
                "bytecodeSize": {
                    "type": "integer"
                },
                "hasCode": {
                    "type": "boolean"
                },
                "network": {
                    "type": "string"
                }
            }
        },
        "contract.ERC20MetadataResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "blockTag": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "totalSupply": {
                    "type": "string"
                }
            }
        },
        "contract.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}